Added my note for archival purposes.
Originally, when I first started writing the script, I wrote other notes and comments alongside it.
Once I switched to writing the script more officially, I copied all of that stuff out into it's own vscode tab.
Since that point onward, there's always been an extra unsaved tab open holding it and other things.
I recently copied my TODO list out of the main script and into the bottom of the note.
Right above that is the very first method I wrote for generating dice. It makes a D6.
The process and code has changed substantially, but the idea is still the exact same. The methods for the other dice were deleted, I only kept one as an example if I needed to copy any logic.
Above that is the code I was writing for the D10 and D12 before I decided to rewrite the script from scratch. Again those got posted there in case I needed to reference them later.
Above that is my little plan for how to spend my time cleaning. I don't know when it was from, but probably from when I was going through my cleaning phase... months ago now.
Above that is the rest of the leftovers from my 2nd version of the script. Over time parts of it were deleted as I didn't need the logic for referencing anymore.
You can figure out the rest I guess. Gonna run madifont2 and graffiti through Blender, then go back to SG.