Added a real Fonts and Dice namespace.
All the dice definitions were moved into the DiceGen::Dice module, and all the fonts were moved into a DiceGen::Fonts module.
The fonts were also split up, so each font has it's own definition file, just like with the die definitions.This was done to make the process for making new fonts and dice symmetric, and more flexible.
Now just dropping a new font into the FontDefinitions folder adds a font, making it almost like a plugin system.
The original reason was that doing it this way will also make IRB list all the fonts and dice models present when it loads, so you have a readymade list of the possible arguments for 'font' and 'model' now.